requirements.txt
flask==2.3.3
flask-bcrypt==1.0.1
flask-cors==4.0.0
mysql-connector-python==9.4.0
numpy
pandas
scikit-learn
xgboost
joblib
gunicorn==20.1.0
